+++ app/uk/gov/hmrc/tai/model/domain/calculation/CodingComponent.scala
@@ -16,7 +16,10 @@
 
 package uk.gov.hmrc.tai.model.domain.calculation
 
-import uk.gov.hmrc.tai.model.domain.TaxComponentType
+import play.api.libs.json.{JsPath, Writes}
+import play.api.libs.functional.syntax._
+import uk.gov.hmrc.tai.model.domain.TaxComponentType.codingComponentTypeWrites
+import uk.gov.hmrc.tai.model.domain.{AllowanceComponentType, BenefitComponentType, DeductionComponentType, NonTaxCodeIncomeComponentType, TaxComponentType}
 
 case class CodingComponent(
   componentType: TaxComponentType,
@@ -25,3 +28,27 @@ case class CodingComponent(
   description: String,
   inputAmount: Option[BigDecimal] = None
 )
+
+object CodingComponent {
+  val codingComponentWrites: Writes[CodingComponent] = (
+    (JsPath \ "componentType").write[TaxComponentType](codingComponentTypeWrites) and
+      (JsPath \ "employmentId").writeNullable[Int] and
+      (JsPath \ "amount").write[BigDecimal] and
+      (JsPath \ "description").write[String] and
+      (JsPath \ "iabdCategory").write[String] and
+      (JsPath \ "inputAmount").writeNullable[BigDecimal]
+    )(unapplyCodingComponentForApiJson _)
+
+  private def unapplyCodingComponentForApiJson(
+                                                t: CodingComponent
+                                              ): (TaxComponentType, Option[Int], BigDecimal, String, String, Option[BigDecimal]) = {
+    val iabdCategory = t.componentType match {
+      case _: AllowanceComponentType        => "Allowance"
+      case _: BenefitComponentType          => "Benefit"
+      case _: DeductionComponentType        => "Deduction"
+      case _: NonTaxCodeIncomeComponentType => "NonTaxCodeIncome"
+      case _                                => throw new RuntimeException("Unrecognised coding Component type")
+    }
+    (t.componentType, t.employmentId, t.amount, t.description, iabdCategory, t.inputAmount)
+  }
+}
